update: remove unneeded injections to right modules
