+++ LOLChampions/LOLChampions/DI /AppDiIntilizer.swift
@@ -10,25 +10,19 @@ import Swinject
 import Home
 import ChampionDetails
 import Common
+import NetworkManger
 import SwiftUI
+import AppCoordinator
 
 class AppDiIntilizer {
     private let container = Container()
     
-    init() {
-        updateNavigationBarColor() 
-    }
-    
     func initDependencies() {
         AppResolver.shared.setDependencyContainer(container)
+        NetwrokDIIntilizer(container: container).buildContainer()
+        AppCoordinatorDIIntilizer(container: container).buildContainer()
         HomeDIIntilizer(container: container).buildContainer()
         ChampionDetailsDIIntilizer(container: container).buildContainer()
-    }
-    
-    private func updateNavigationBarColor() {
-        let navigationColor = UIColor(Color(red: 0.067, green: 0.098, blue: 0.133))
-        UINavigationBar.appearance().barTintColor = navigationColor
-        UINavigationBar.appearance().backgroundColor = navigationColor
     }
     
 }
