+++ LOLChampions/LOLChampions/LOLChampionsApp.swift
@@ -10,6 +10,7 @@ import SwiftUI
 @main
 struct LOLChampionsApp: App {
     init() {
+        updateNavigationBarColor()
         AppDiIntilizer().initDependencies()
     }
     
@@ -17,5 +18,11 @@ struct LOLChampionsApp: App {
         WindowGroup {
             ContentView()
         }
+    }
+    
+    private func updateNavigationBarColor() {
+        let navigationColor = UIColor(Color(red: 0.067, green: 0.098, blue: 0.133))
+        UINavigationBar.appearance().barTintColor = navigationColor
+        UINavigationBar.appearance().backgroundColor = navigationColor
     }
 }
