+++ features/ChampionDetails/ChampionDetailsTests/ChampionInfoUseCaseTests.swift
@@ -35,9 +35,7 @@ extension ChampionInfoUseCaseTests {
     func initDI() {
         let container = Container()
         AppResolver.shared.setDependencyContainer(container)
-        container.register(NetworkClientType.self) { _  in
-            return NetworkClient()
-        }
+        NetwrokDIIntilizer(container: container).buildContainer()
 
         container.register(ChampionInfoServiceType.self) { _  in
             return ChampionInfoServiceMock()
